% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_criterion.R
\name{criterion_ocean_uptake_obs}
\alias{criterion_ocean_uptake_obs}
\title{Screening criterion using observed ocean carbon sink}
\usage{
criterion_ocean_uptake_obs()
}
\value{
A criterion identifier using observed ocean carbon sink
#' @note This function uses data from The Global Carbon Project's 2023 Global Carbon Budget. Friedlingstein et al. Global Carbon Budget 2023, Earth Sys. Sci. Data, 15, 5301-5369. It can be accessed here:
\href{https://globalcarbonbudgetdata.org/latest-data.html}
}
\description{
This is a criterion identifier for screening model runs using
observed ocean carbon sink values from The Global Carbon Project data set.
}
\examples{
criterion_ocean_uptake_obs()
}
